+++ src/main/java/vendingmachine/VendingMachine.java
@@ -1,10 +1,12 @@
 package vendingmachine;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
-import java.util.List;
 
+import vendingmachine.validator.InputAmountValidator;
+import vendingmachine.validator.InputBuyItemValidator;
+import vendingmachine.validator.InputItemValidator;
+import vendingmachine.validator.InputUserAmountValidator;
 import vendingmachine.view.InputView;
 import vendingmachine.view.OutputView;
