+++ src/main/java/vendingmachine/validator/InputAmountValidator.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.validator;
 
 public class InputAmountValidator {
 	private final static String ONLY_NUMBER = "[0-9]+";
