+++ src/main/java/vendingmachine/validator/InputBuyItemValidator.java
@@ -1,6 +1,8 @@
-package vendingmachine;
+package vendingmachine.validator;
 
 import java.util.HashMap;
+
+import vendingmachine.Item;
 
 public class InputBuyItemValidator {
 	private final static String EXIST_ERROR_MESSAGE = "[ERROR] 구매하려는 상품이 존재하지 않습니다.";
